+++ src/AddTodoForm.js
@@ -1,0 +1,20 @@
+import React from "react";
+
+
+function AddTodoForm(){
+
+    return(
+        <form > 
+            <label htmlFor="todoTitle">
+                Title
+            </label>
+            <input type="text" id="todoTitle" name="title"/> 
+            <button type="submit">
+                Add
+            </button>
+        </form>
+    )
+
+}
+
+export default AddTodoForm;
